+++ NEWS.md
@@ -20,6 +20,7 @@
   * `bs_remove_reaction()` removes a reaction to a message
   * `bs_mute_convo()` mutes a conversation
   * `bs_unmute_convo()` unmutes a conversation
+  * `bs_send_message()` sends a message to a conversation
 
 # bskyr 0.3.0